refactor finetuner calc to put supporting functions and classes into different files
