+++ finetuna/ml_potentials/finetuner_utils/__init__.py
@@ -1,0 +1,1 @@
+from .utils import GemNetTUncertainty
